# docs-site/static/openapi/interakt-v1.yaml
openapi: 3.1.0
info:
  title: Interakt Public API
  version: 1.0.0
  description: >-
    Public, token-authenticated REST API for embedding Interakt search and AI
    chat into your own applications.


    All endpoints live under `/api/v1`. Authenticate with your experience’s
    access token via the `X-Access-Token` header (or `Authorization: Bearer`).
    Document ingestion uses a separate per-index `X-Api-Key`.
servers:
  - url: https://admin.interakt.app
    description: Hosted Interakt
...
components:
  securitySchemes:
    AccessToken:
      type: apiKey
      in: header
      name: X-Access-Token
      description: Per-experience access token. Find it in the dashboard under the
        Search or AI Experience’s "Access" tab.
    BearerAuth:
      type: http
      scheme: bearer
      description: "The same access token may be sent as `Authorization: Bearer
        <token>` instead of `X-Access-Token`."
    IngestApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Per-index ingestion API key, used only by the document ingestion
        endpoint.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
          example: Invalid access token
        code:
          type: string
          example: UNAUTHORIZED
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
            required:
              - field
              - message
          description: Per-field validation errors (present on 400 VALIDATION_ERROR).
      required:
        - error
...
    ChatRequest:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 8000
        sessionId:
          type: string
          maxLength: 200
      required:
        - message
...
paths:
...
  /api/v1/ai-experiences/chat:
    post:
      tags:
        - AI
      summary: Chat with an AI Experience (streaming)
      description: "Sends a user message to the AI Experience identified by the access
        token and streams the assistant response. Server-Sent Events stream
        (`text/event-stream`). Each line is `data: <json>`. Event objects carry
        a `type` of `content` | `tool_call` | `tool_result` | `done` | `error`."
      security: *a1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: SSE stream of the assistant turn.
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: AI Experience is not active.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...